// apps/web/src/api/generated/models/notificationDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * APU Automotive Service Centre (APU-ASC) API
 * Automotive Service Management RESTful API Specifications
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationDtoType } from './notificationDtoType'

export interface NotificationDto {
  id?: string
  type?: NotificationDtoType
  title?: string
  body?: string
  link?: string
  readAt?: string
  createdAt?: string
}